Programming Bot/chatbot-flask-main/requirements.txt
Flask==3.0.3
ollama==0.3.3
llama-index
llama-index-embeddings-huggingface==0.4.0
llama-index-llms-ollama==0.4.0
transformers==4.46.2
faiss-cpu==1.9.0.post1
numpy==1.26.4